+++ src/Record/tests.py
@@ -29,6 +29,14 @@ class R(Record):
     __record_schema__ = {'a': 0, 'b': 1, 'c': 2}
 
 
+class RSameSchema(Record):
+    __record_schema__ = {'a': 0, 'b': 1, 'c': 2}
+
+
+class RDifferentSchema(Record):
+    __record_schema__ = {'x': 0, 'y': 1, 'z': 2}
+
+
 class RecordTest(unittest.TestCase):
 
     def test_init(self):
@@ -161,6 +169,33 @@ class RecordTest(unittest.TestCase):
     def test_len(self):
         r = R((1, 2, None))
         self.assertEqual(len(r), 3)
+
+    def test_hash(self):
+        r1 = R((1, 2, None))
+        r2 = R((1, 2, None))
+        self.assertNotEqual(hash(r1), hash(r2))
+
+    def test_hash_schema(self):
+        r = R((1, 2, None))
+        r_same = RSameSchema((1, 2, None))
+        r_diff = RDifferentSchema((1, 2, None))
+        self.assertNotEqual(hash(r), hash(r_same))
+        self.assertNotEqual(hash(r), hash(r_diff))
+
+    def test_set_members(self):
+        r1 = R((1, 2, None))
+        r2 = R((1, 2, None))
+        r3 = R((1, 2, None))
+        r_diff = RDifferentSchema((1, 2, None))
+        records = set([r1])
+        records.add(r2)
+        self.assertTrue(r1 in records)
+        self.assertTrue(r2 in records)
+        self.assertFalse(r3 in records)
+        self.assertEqual(len(records), 2)
+        records.add(r_diff)
+        self.assertEqual(len(records), 3)
+        self.assertTrue(r_diff in records)
 
     def test_cmp(self):
         r1 = R((1, 2, 0))